Fix conversion bug for JSON parsing on 32bit devices
Better 32bit architecture support.

Changed the parsing logic to convert Any type values to Double first, before converting to CGFloat.  Casting CGFloat's from the JSON values fail, causing annotations to not render.

+++ Pollock/Data Structure/Location.swift
@@ -18,14 +18,14 @@ internal struct Location : Serializable {
     }
 
     init(_ payload: [String : Any]) throws {
-        guard let xOffset = payload["xOffset"] as? CGFloat else {
+        guard let xOffset = payload["xOffset"] as? Double else {
             throw SerializerError("Missing xOffset for Location")
         }
-        guard let yOffset = payload["yOffset"] as? CGFloat else {
+        guard let yOffset = payload["yOffset"] as? Double else {
             throw SerializerError("Missing yOffset for Location")
         }
-        self.xOffset = xOffset
-        self.yOffset = yOffset
+        self.xOffset = CGFloat(xOffset)
+        self.yOffset = CGFloat(yOffset)
     }
 
     func serialize() throws -> [String : Any] {
